backend/[1;33m[WARNING][0m Directory 'CLAUDE_AINAK' already exists. Cleaning it...[0;32m[SUCCESS][0m Created directory: CLAUDE_AINAKCLAUDE_AINAK/backend/requirements.txt
# LFA Legacy GO - Core Backend Dependencies
# Cleaned version - only essential packages

# FastAPI Framework
fastapi==0.116.1
uvicorn==0.35.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.7

# Authentication & Security  
pydantic==2.11.7
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==4.0.1

# Development
pytest==8.3.2

# Utilities
python-dotenv==1.0.1
requests==2.32.3